Add support for other editors
Right now, the language server makes an RPC call to get the contents of
the buffers open in the editor when it needs to access them. This
doesn't work with servers that don't happen to be running the VS Code
plugin.

The standard way to do this is for the language server to keep a copy of
all the open files in memory and refer to that when necessary. This
commit does this, which allows it to work in Facebook's Nuclide, and no
doubt other editors in due time.

I tested it in Nuclide and got things like types on hover to show up in
an OCaml source files, which didn't work before this commit.