Fix autocomplete
Autocomplete wasn't working in Nuclide before I added support for other
editors, so I wasn't confused when it wasn't working afterwards. I
completely misinterpreted what `getPrefix` was supposed to do and just
returned from the entire beginning of the line, but really we only want
from the beginning of this expression. Autocomplete actually works in
Nuclide now (and should probably not be broken in other editors, but I
didn't figure out how to test the extension in VS Code).